feat: improve file structure
